+++ src/VisaServiceProvider.php
@@ -2,9 +2,13 @@
 
 namespace Stratedge\Visa;
 
+use DateInterval;
+use Laravel\Passport\Bridge\PersonalAccessGrant;
 use Laravel\Passport\Client as BaseClient;
 use Laravel\Passport\Passport;
 use Laravel\Passport\PassportServiceProvider as BaseServiceProvider;
+use League\OAuth2\Server\AuthorizationServer;
+use League\OAuth2\Server\Grant\ClientCredentialsGrant;
 
 class VisaServiceProvider extends BaseServiceProvider
 {
@@ -105,5 +109,63 @@ class VisaServiceProvider extends BaseServiceProvider
             \Laravel\Passport\Http\Controllers\AuthorizationController::class,
             \Stratedge\Visa\Http\Controllers\AuthorizationController::class
         );
+    }
+
+        /**
+     * Register the authorization server.
+     *
+     * @return void
+     */
+    protected function registerAuthorizationServer()
+    {
+        $this->app->singleton(AuthorizationServer::class, function () {
+            return tap($this->makeAuthorizationServer(), function ($server) {
+                $server->enableGrantType(
+                    $this->makeAuthCodeGrant(),
+                    Passport::tokensExpireIn()
+                );
+
+                $server->enableGrantType(
+                    $this->makeRefreshTokenGrant(),
+                    Passport::tokensExpireIn()
+                );
+
+                $server->enableGrantType(
+                    $this->makePasswordGrant(),
+                    Passport::tokensExpireIn()
+                );
+
+                $server->enableGrantType(
+                    new PersonalAccessGrant,
+                    new DateInterval('P1Y')
+                );
+
+                $server->enableGrantType(
+                    new ClientCredentialsGrant,
+                    Passport::tokensExpireIn()
+                );
+
+                if (Passport::$implicitGrantEnabled) {
+                    $server->enableGrantType(
+                        $this->makeImplicitGrant(),
+                        Passport::tokensExpireIn()
+                    );
+                }
+
+                //Add additional grants through easily overloaded method
+                $this->enableCustomGrants($server);
+            });
+        });
+    }
+
+    /**
+     * Enables any additional custom grants when overloaded.
+     *
+     * @param  AuthorizationServer $server
+     * @return void
+     */
+    public function enableCustomGrants(AuthorizationServer $server)
+    {
+        //
     }
 }